architecture/agents/requirements.txt
FlightRadarAPI
azure-cosmos
python-dotenv